Enc common: removing decoder build dependencies
