+++ Source/Lib/Common/ASM_AVX2/aom_subpixel_8t_intrin_avx2.c
@@ -11,7 +11,7 @@
 
 #include <immintrin.h>
 
-#include "aom_dsp_rtcd.h"
+#include "common_dsp_rtcd.h"
 
 #include "convolve.h"
 #include "convolve_avx2.h"
