+++ Source/Lib/Common/Codec/EbSystemResourceManager.h
