+++ Source/Lib/Decoder/CMakeLists.txt
@@ -19,16 +19,8 @@ include_directories(${PROJECT_SOURCE_DIR}/Source/API/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Common/ASM_SSE4_1/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Common/ASM_AVX2/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Common/ASM_AVX512/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/Codec/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/C_DEFAULT/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSE2/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSSE3/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSE4_1/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_AVX2/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_AVX512/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Decoder/Codec/
-    ${PROJECT_SOURCE_DIR}/third_party/fastfeat/
-        )
+    ${PROJECT_SOURCE_DIR}/third_party/fastfeat/)
 
 link_directories(
      ${PROJECT_SOURCE_DIR}/Source/Lib/Common/ASM_SSE2/
@@ -39,15 +31,7 @@ link_directories(
     ${PROJECT_SOURCE_DIR}/Source/Lib/Common/ASM_AVX512/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Common/Codec/
     ${PROJECT_SOURCE_DIR}/Source/Lib/Decoder/Codec/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/Codec/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/C_DEFAULT/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSE2/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSSE3/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_SSE4_1/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_AVX2/
-    ${PROJECT_SOURCE_DIR}/Source/Lib/Encoder/ASM_AVX512/
-    ${PROJECT_SOURCE_DIR}/third_party/fastfeat/
-)
+    ${PROJECT_SOURCE_DIR}/third_party/fastfeat/)
 
 file(GLOB all_files
     "Codec/*.h"
@@ -81,14 +65,7 @@ add_library(SvtAv1Dec
     $<TARGET_OBJECTS:COMMON_ASM_SSSE3>
     $<TARGET_OBJECTS:COMMON_ASM_SSE4_1>
     $<TARGET_OBJECTS:COMMON_ASM_AVX2>
-    $<TARGET_OBJECTS:COMMON_ASM_AVX512>
-     $<TARGET_OBJECTS:ENCODER_C_DEFAULT>
-     $<TARGET_OBJECTS:ENCODER_ASM_SSE2>
-     $<TARGET_OBJECTS:ENCODER_ASM_SSSE3>
-     $<TARGET_OBJECTS:ENCODER_ASM_SSE4_1>
-     $<TARGET_OBJECTS:ENCODER_ASM_AVX2>
-     $<TARGET_OBJECTS:ENCODER_ASM_AVX512>
-     $<TARGET_OBJECTS:ENCODER_CODEC>)
+    $<TARGET_OBJECTS:COMMON_ASM_AVX512>)
 
 set_target_properties(SvtAv1Dec PROPERTIES VERSION ${DEC_VERSION})
 set_target_properties(SvtAv1Dec PROPERTIES SOVERSION ${DEC_VERSION_MAJOR})
